Added code to activate HMR. Messages and routing code are now in separate files.
